fix: league mockup style
