+++ components/league/dndSortableItem.tsx
@@ -13,7 +13,7 @@ const SortableItem = ({ id }: { id: UniqueIdentifier }) => {
             {...attributes}
             {...listeners}
         >
-            <Card sx={{p:2, backgroundColor:"primary.main", color:"secondary.main", minWidth:"200px", maxWidth:"200px"}}>
+            <Card sx={{p:1.5, backgroundColor:"primary.dark", color:"secondary.main", width:"130px", overflow:"hidden"}}>
                 <Item id={id} />
             </Card>
         </div>
